requirements.txt
Django==1.11.11
django-cors-headers==2.2.0
django-filter==1.1.0
django-rest-framework-mongoengine==3.3.1
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
Markdown==2.6.11
mongoengine==0.15.0
PyJWT==1.6.0
pymongo==3.6.1
pytz==2018.3
six==1.11.0
